+++ packages/module-social/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @m5kdev/module-social
 
+## 0.36.0
+
+### Minor Changes
+
+- 0.36.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @m5kdev/backend@0.36.0
+
 ## 0.35.0
 
 ### Minor Changes