+++ packages/web-ui/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @m5kdev/web-ui
 
+## 0.36.0
+
+### Minor Changes
+
+- 0.36.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @m5kdev/commons@0.36.0
+  - @m5kdev/frontend@0.36.0
+
 ## 0.35.0
 
 ### Minor Changes